Fix issue of high stats value being sent from VN UVE.
Issue
VN stats are picked from flow statistics.

Statistics for VN are collected from flow statistics. Agent periodically collects flow stats from vrouter and updates its statistics for flow.
During such collection agent also detects whether 6 byte statistics provided by vrouter has resulted in overflow. If it detects overflow, it will increment stats by value of 0x0001000000000000 (first bit of 7th byte) to account for overflow. Because of bug in overflow detection, the stats were getting incremented by 0x0001000000000000, even when overflow has not really occurred. This is resulting in high VN stats.
 
During periodic stats collection from vrouter, agent detects overflow by checking if the value of agent copy of stats is bigger than the value of vrouter copy of stats. The agent copy of stats was higher than vrouter copy due to the following reasons
1.  When gen-id/flow-handle of a flow changes, statistics for the flow were not reset by stats collection module.
2.  During periodic scanning of flows, the stats collection module used current/updated flow-handle and gen-id from the flow.
3.  When vrouter sends message for stats collection, agent was not using the gen-id in the message.

Fix
1.When stats collection module receives add for a flow which is already present, reset stats for the flow if its gen-id/flow-handle (identified by change of flow uuid) has changed.
2. During periodic scanning of flows, use gen-id/flow-handle with which the flow was added to stats collection module, instead of using the latest gen-id/flow-handle of flow.
3. The Evicted flows stats update message sent to stats collection module should carry UUID of the flow. In Stats collection module ignore the stats update message if the Flow UUID in the message does not match with its Flow UUID.
4. Read all the stats, flags, udp source port for flows in one shot and validate the read values. Use these read values for further processing instead of reading these fields from vrouter separately during processing.
5. When stats collection module is processing delete message for a flow, we read and export stats for the flow before it is deleted. To read stats here we should use gen-id/flow-handle from flow stats module instead of picking latest value from FlowEntry.

Closes-Bug: #1654557

+++ src/vnsw/agent/vrouter/ksync/flowtable_ksync.h
@@ -28,7 +28,7 @@ class KSyncFlowIndexManager;
 struct FlowKSyncResponseInfo {
     int ksync_error_;
     uint32_t flow_handle_;
-    uint32_t gen_id_;
+    uint8_t gen_id_;
     uint64_t evict_flow_bytes_;
     uint64_t evict_flow_packets_;
     int32_t evict_flow_oflow_;
@@ -92,7 +92,7 @@ public:
         ksync_response_info_.Reset();
     }
     void SetKSyncResponseInfo(int ksync_error, uint32_t flow_handle,
-                              uint32_t gen_id, uint64_t evict_flow_bytes,
+                              uint8_t gen_id, uint64_t evict_flow_bytes,
                               uint64_t evict_flow_packets,
                               int32_t evict_flow_oflow) {
         ksync_response_info_.ksync_error_ = ksync_error;
